refactor: move out bootloaders in modules
