+++ composer.json
@@ -20,13 +20,13 @@
     "ext-curl": "*",
     "ctx/module-research": "^1.0",
     "ctx/module-config-templates": "^1.0",
-    "ctx/mcp-server": "^1.1.3",
+    "ctx/mcp-server": "^1.1.4",
     "league/html-to-markdown": "^5.1",
     "psr/log": "^3.0",
     "symfony/finder": "^6.0 | ^7.0 | ^8.0",
     "symfony/console": "^6.0 | ^7.0 | ^8.0",
     "symfony/process": "^6.0 | ^7.0 | ^8.0",
-    "symfony/yaml": "^7.2",
+    "symfony/yaml": "^6.0 | ^7.0 | ^8.0",
     "nette/php-generator": "^4.1",
     "nikic/php-parser": "^v4.0 | ^v5.0",
     "vlucas/phpdotenv": "^5.6",
