+++ phpunit.xml
@@ -15,6 +15,12 @@
         </testsuite>
     </testsuites>
     <php>
+        <env name="OAUTH_ENABLED" value="false" />
+        <env name="OAUTH_ISSUER_URL" value="http://127.0.0.1:8090" />
+        <env name="OAUTH_SERVER_URL" value="http://127.0.0.1:8090" />
+        <env name="OAUTH_CLIENT_ID" value="test-id" />
+        <env name="OAUTH_CLIENT_SECRET" value="test-secret" />
+
         <ini name="error_reporting" value="-1"/>
         <ini name="memory_limit" value="-1"/>
     </php>
