+++ src/Application/Kernel.php
@@ -4,21 +4,21 @@ declare(strict_types=1);
 
 namespace Butschster\ContextGenerator\Application;
 
-use Butschster\ContextGenerator\Application\Bootloader\ComposerClientBootloader;
+use Butschster\ContextGenerator\Lib\ComposerClient\ComposerClientBootloader;
 use Butschster\ContextGenerator\Config\ConfigLoaderBootloader;
 use Butschster\ContextGenerator\Application\Bootloader\ConfigurationBootloader;
 use Butschster\ContextGenerator\Application\Bootloader\ConsoleBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\ContentRendererBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\Context7ClientBootloader;
+use Butschster\ContextGenerator\Lib\Content\ContentRendererBootloader;
+use Butschster\ContextGenerator\Lib\Context7Client\Context7ClientBootloader;
 use Butschster\ContextGenerator\Application\Bootloader\CoreBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\ExcludeBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\GithubClientBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\GitlabClientBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\HttpClientBootloader;
+use Butschster\ContextGenerator\Config\Exclude\ExcludeBootloader;
+use Butschster\ContextGenerator\Lib\GithubClient\GithubClientBootloader;
+use Butschster\ContextGenerator\Lib\GitlabClient\GitlabClientBootloader;
+use Butschster\ContextGenerator\Lib\HttpClient\HttpClientBootloader;
 use Butschster\ContextGenerator\Application\Bootloader\LoggerBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\ModifierBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\SourceFetcherBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\VariableBootloader;
+use Butschster\ContextGenerator\Modifier\ModifierBootloader;
+use Butschster\ContextGenerator\Source\Fetcher\SourceFetcherBootloader;
+use Butschster\ContextGenerator\Lib\Variable\VariableBootloader;
 use Butschster\ContextGenerator\Research\ResearchBootloader;
 use Butschster\ContextGenerator\McpServer\ActionsBootloader;
 use Butschster\ContextGenerator\Template\TemplateSystemBootloader;
