+++ src/Config/Exclude/ExcludeBootloader.php
@@ -2,12 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Butschster\ContextGenerator\Application\Bootloader;
+namespace Butschster\ContextGenerator\Config\Exclude;
 
 use Butschster\ContextGenerator\Config\ConfigLoaderBootloader;
-use Butschster\ContextGenerator\Config\Exclude\ExcludeParserPlugin;
-use Butschster\ContextGenerator\Config\Exclude\ExcludeRegistry;
-use Butschster\ContextGenerator\Config\Exclude\ExcludeRegistryInterface;
 use Spiral\Boot\Bootloader\Bootloader;
 use Spiral\Core\Attribute\Singleton;
