+++ src/Lib/ComposerClient/ComposerClientBootloader.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Butschster\ContextGenerator\Lib\ComposerClient;
+
+use Spiral\Boot\Bootloader\Bootloader;
+
+final class ComposerClientBootloader extends Bootloader
+{
+    #[\Override]
+    public function defineSingletons(): array
+    {
+        return [
+            ComposerClientInterface::class => FileSystemComposerClient::class,
+        ];
+    }
+}
