+++ src/Lib/Variable/VariableBootloader.php
@@ -2,20 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Butschster\ContextGenerator\Application\Bootloader;
+namespace Butschster\ContextGenerator\Lib\Variable;
 
 use Butschster\ContextGenerator\Config\ConfigLoaderBootloader;
 use Butschster\ContextGenerator\Config\Parser\VariablesParserPlugin;
 use Butschster\ContextGenerator\DirectoriesInterface;
-use Butschster\ContextGenerator\Lib\Variable\CompositeProcessor;
 use Butschster\ContextGenerator\Lib\Variable\Provider\CompositeVariableProvider;
 use Butschster\ContextGenerator\Lib\Variable\Provider\ConfigVariableProvider;
 use Butschster\ContextGenerator\Lib\Variable\Provider\DotEnvVariableProvider;
 use Butschster\ContextGenerator\Lib\Variable\Provider\PredefinedVariableProvider;
 use Butschster\ContextGenerator\Lib\Variable\Provider\VariableProviderInterface;
-use Butschster\ContextGenerator\Lib\Variable\VariableReplacementProcessor;
-use Butschster\ContextGenerator\Lib\Variable\VariableReplacementProcessorInterface;
-use Butschster\ContextGenerator\Lib\Variable\VariableResolver;
 use Dotenv\Repository\RepositoryBuilder;
 use Spiral\Boot\Bootloader\Bootloader;
