+++ src/Modifier/ModifierBootloader.php
@@ -2,11 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Butschster\ContextGenerator\Application\Bootloader;
+namespace Butschster\ContextGenerator\Modifier;
 
 use Butschster\ContextGenerator\Config\ConfigLoaderBootloader;
 use Butschster\ContextGenerator\Modifier\Alias\ModifierAliasesParserPlugin;
-use Butschster\ContextGenerator\Modifier\SourceModifierRegistry;
 use Spiral\Boot\Bootloader\Bootloader;
 
 final class ModifierBootloader extends Bootloader
