+++ src/Source/Composer/ComposerSourceBootloader.php
@@ -4,9 +4,13 @@ declare(strict_types=1);
 
 namespace Butschster\ContextGenerator\Source\Composer;
 
-use Butschster\ContextGenerator\Application\Bootloader\ComposerClientBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\SourceFetcherBootloader;
+use Butschster\ContextGenerator\Lib\ComposerClient\ComposerClientBootloader;
+use Butschster\ContextGenerator\Source\Fetcher\SourceFetcherBootloader;
+use Butschster\ContextGenerator\Application\Logger\HasPrefixLoggerInterface;
 use Butschster\ContextGenerator\DirectoriesInterface;
+use Butschster\ContextGenerator\Source\Composer\Provider\ComposerProviderInterface;
+use Butschster\ContextGenerator\Source\Composer\Provider\CompositeComposerProvider;
+use Butschster\ContextGenerator\Source\Composer\Provider\LocalComposerProvider;
 use Butschster\ContextGenerator\Source\Registry\SourceRegistryInterface;
 use Spiral\Boot\Bootloader\Bootloader;
 use Spiral\Core\FactoryInterface;
@@ -23,6 +27,14 @@ final class ComposerSourceBootloader extends Bootloader
     public function defineSingletons(): array
     {
         return [
+            ComposerProviderInterface::class => static fn(
+                HasPrefixLoggerInterface $logger,
+                LocalComposerProvider $localProvider,
+            ) => new CompositeComposerProvider(
+                logger: $logger->withPrefix('composer-provider'),
+                localProvider: $localProvider,
+            ),
+
             ComposerSourceFetcher::class => static fn(
                 FactoryInterface $factory,
                 DirectoriesInterface $dirs,
