+++ src/Source/Docs/DocsSourceBootloader.php
@@ -4,12 +4,8 @@ declare(strict_types=1);
 
 namespace Butschster\ContextGenerator\Source\Docs;
 
-use Butschster\ContextGenerator\Application\Bootloader\HttpClientBootloader;
-use Butschster\ContextGenerator\Application\Bootloader\SourceFetcherBootloader;
-use Butschster\ContextGenerator\Application\Logger\HasPrefixLoggerInterface;
-use Butschster\ContextGenerator\Lib\Content\ContentBuilderFactory;
-use Butschster\ContextGenerator\Lib\HttpClient\HttpClientInterface;
-use Butschster\ContextGenerator\Lib\Variable\VariableResolver;
+use Butschster\ContextGenerator\Lib\HttpClient\HttpClientBootloader;
+use Butschster\ContextGenerator\Source\Fetcher\SourceFetcherBootloader;
 use Butschster\ContextGenerator\Source\Registry\SourceRegistryInterface;
 use Spiral\Boot\Bootloader\Bootloader;
 use Spiral\Core\FactoryInterface;
@@ -28,10 +24,6 @@ final class DocsSourceBootloader extends Bootloader
         return [
             DocsSourceFetcher::class => static fn(
                 FactoryInterface $factory,
-                ContentBuilderFactory $builderFactory,
-                HttpClientInterface $httpClient,
-                VariableResolver $variables,
-                HasPrefixLoggerInterface $logger,
             ): DocsSourceFetcher => $factory->make(DocsSourceFetcher::class, [
                 'defaultHeaders' => [
                     'User-Agent' => 'CTX Bot',
